Persist terminal session across rerenders
