+++ src/components/terminal/Terminals.tsx
@@ -15,11 +15,10 @@ import {
   IonToolbar,
 } from '@ionic/react';
 import { close, ellipsisHorizontal, ellipsisVertical } from 'ionicons/icons';
-import React, { useContext, useState } from 'react';
+import React, { useState } from 'react';
 
-import { IContext, ITerminal } from '../../declarations';
-import { AppContext } from '../../utils/context';
-import Terminal from './Terminal';
+import { ITerminal } from '../../declarations';
+import Shell from './Shell';
 
 interface ITerminalsProps {
   showModal: boolean;
@@ -38,8 +37,6 @@ const Terminals: React.FunctionComponent<ITerminalsProps> = ({
   setActiveTerminal,
   removeTerminal,
 }: ITerminalsProps) => {
-  const context = useContext<IContext>(AppContext);
-
   const [showPopover, setShowPopover] = useState<boolean>(false);
   const [popoverEvent, setPopoverEvent] = useState();
 
@@ -85,7 +82,7 @@ const Terminals: React.FunctionComponent<ITerminalsProps> = ({
                       removeTerminal(index);
                     }}
                   >
-                    <IonLabel>{`Close ${terminal.container}`}</IonLabel>
+                    <IonLabel>{`Close ${terminal.name}`}</IonLabel>
                   </IonItem>
                 );
               })}
@@ -103,7 +100,7 @@ const Terminals: React.FunctionComponent<ITerminalsProps> = ({
             {terminals.map((terminal, index) => {
               return (
                 <IonSegmentButton key={index} value={`term_${index}`}>
-                  <IonLabel>{terminal.container}</IonLabel>
+                  <IonLabel>{terminal.name}</IonLabel>
                 </IonSegmentButton>
               );
             })}
@@ -113,9 +110,7 @@ const Terminals: React.FunctionComponent<ITerminalsProps> = ({
 
       <IonContent>
         {terminals.map((terminal, index) => {
-          return activeTerminal === `term_${index}` ? (
-            <Terminal key={index} terminal={terminal} darkMode={context.settings.darkMode} />
-          ) : null;
+          return activeTerminal === `term_${index}` ? <Shell key={index} terminal={terminal} /> : null;
         })}
       </IonContent>
     </IonModal>
